Move service logic to presenter
